separate feathers migration code to separate helper file
